# demo/message-center-demo/contracts/openapi/message-center-v1.yaml
openapi: 3.1.0
info:
  title: CRM for Logistics Message Center API
  version: 1.0.0
  description: Versioned HTTP contract for the message center frontend and backend.
  contact:
    name: CRM for Logistics Engineering
servers:
  - url: /
...
security:
  - sessionCookie: []
paths:
...
  /api/v1/conversations/{conversationId}/messages:
    get:
      operationId: listMessages
      description: List messages from an authorized conversation using bounded cursor pagination.
      tags:
        - Messaging
      parameters:
        - $ref: '#/components/parameters/ConversationId'
        - $ref: '#/components/parameters/TraceRequestId'
        - $ref: '#/components/parameters/CursorParameter'
        - $ref: '#/components/parameters/LimitParameter'
      responses:
        '200':
          description: Authorized conversation message page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagePage'
        '4XX':
          $ref: '#/components/responses/ApiErrorResponse'
        default:
          $ref: '#/components/responses/ApiErrorResponse'
    post:
      operationId: sendMessage
      description: Atomically queue a pending outbound message without calling a channel directly.
      tags:
        - Messaging
      parameters:
        - $ref: '#/components/parameters/ConversationId'
        - $ref: '#/components/parameters/TraceRequestId'
        - $ref: '#/components/parameters/CsrfTokenHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '202':
          description: Pending message and outbox job committed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueuedMessage'
        '4XX':
          $ref: '#/components/responses/ApiErrorResponse'
        default:
          $ref: '#/components/responses/ApiErrorResponse'
...
  /api/v1/email/messages:
    post:
      operationId: sendEmailMessage
      description: Send one UTF-8 plain-text email as JSON or multipart form data with zero to sixteen ordinary attachments. Attachment content is bounded to 20 MiB in total. If SMTP delivery may have succeeded but local persistence is incomplete, return 503 EMAIL_SEND_OUTCOME_UNKNOWN; inspect delivery before retrying. Stable errors include EMAIL_ATTACHMENT_COUNT_LIMIT, EMAIL_ATTACHMENT_SIZE_LIMIT, EMAIL_ATTACHMENT_STORAGE_FULL, EMAIL_ATTACHMENT_READ_FAILED, EMAIL_SUBMISSION_PERSIST_FAILED and EMAIL_SEND_OUTCOME_UNKNOWN.
      tags:
        - Messaging
      parameters:
        - $ref: '#/components/parameters/TraceRequestId'
        - $ref: '#/components/parameters/CsrfTokenHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailSendJsonRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/EmailSendMultipartRequest'
      responses:
        '200':
          description: SMTP accepted the message and its local message record was persisted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailSendResult'
        '503':
          description: SMTP outcome is unknown or the accepted message could not be fully recorded locally. Do not retry without checking delivery.
          $ref: '#/components/responses/ApiErrorResponse'
        '4XX':
          $ref: '#/components/responses/ApiErrorResponse'
        default:
          $ref: '#/components/responses/ApiErrorResponse'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: MESSAGE_CENTER_SESSION
      description: Opaque session token. The server stores only its SHA-256 digest.
    webhookSignature:
      type: apiKey
      in: header
      name: X-Webhook-Signature
      description: Header signature scheme. The adapter also requires X-Webhook-Timestamp and X-Webhook-Nonce for applicable channel types and verifies the exact raw body before parsing.
    weComSignature:
      type: apiKey
      in: query
      name: msg_signature
      description: WeCom query signature scheme. The adapter also requires timestamp and nonce query parameters and verifies the exact raw XML body before parsing.
  parameters:
    TraceRequestId:
      name: X-Request-ID
      in: header
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 128
    CsrfTokenHeader:
      name: X-XSRF-TOKEN
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 512
...
    CursorParameter:
      name: cursor
      in: query
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 2048
    LimitParameter:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
        default: 50
...
    ConversationId:
      name: conversationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ApiErrorResponse:
      description: Structured failure with no credential or signed URL material
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
...
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message, traceId, fieldErrors]
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 100
        message:
          type: string
          minLength: 1
          maxLength: 1000
        traceId:
          type: string
          minLength: 1
          maxLength: 128
        fieldErrors:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      additionalProperties: false
      required: [field, code, message]
      properties:
        field:
          type: string
          minLength: 1
          maxLength: 200
        code:
          type: string
          minLength: 1
          maxLength: 100
        message:
          type: string
          minLength: 1
          maxLength: 500
...
    MessagePage:
      type: object
      additionalProperties: false
      required: [items, nextCursor, messageCount, threadRevision]
      properties:
        items:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/Message'
        nextCursor:
          type: [string, 'null']
          maxLength: 2048
        messageCount:
          type: integer
          minimum: 0
        threadRevision:
          type: string
          minLength: 1
          maxLength: 128
    Message:
      type: object
      additionalProperties: false
      required: [id, conversationId, channelAccountId, direction, messageKind,
        occurredAt, currentStatus, ingestSequence, countsAsUnread, attachments]
      properties:
        id:
          type: string
          format: uuid
        conversationId:
          type: string
          format: uuid
        channelAccountId:
          type: string
          format: uuid
        providerMessageId:
          type: [string, 'null']
          maxLength: 255
        clientRequestId:
          type: [string, 'null']
          maxLength: 255
        direction:
          type: string
          enum: [inbound, outbound, system]
        messageKind:
          $ref: '#/components/schemas/MessageKind'
        subject:
          type: [string, 'null']
          maxLength: 1000
        bodyText:
          type: [string, 'null']
          maxLength: 100000
        occurredAt:
          type: string
          format: date-time
        currentStatus:
          $ref: '#/components/schemas/MessageStatus'
        ingestSequence:
          type: integer
          format: int64
          minimum: 1
        countsAsUnread:
          type: boolean
        attachments:
          type: array
          maxItems: 10
          items:
            $ref: '#/components/schemas/Attachment'
    SendMessageRequest:
      type: object
      additionalProperties: false
      required: [channelAccountId, clientRequestId, messageKind, attachmentIds]
      properties:
        channelAccountId:
          type: string
          format: uuid
        clientRequestId:
          type: string
          minLength: 1
          maxLength: 255
        messageKind:
          $ref: '#/components/schemas/MessageKind'
        subject:
          type: [string, 'null']
          maxLength: 1000
        bodyText:
          type: [string, 'null']
          maxLength: 100000
        attachmentIds:
          type: array
          uniqueItems: true
          maxItems: 10
          items:
            type: string
            format: uuid
    QueuedMessage:
      type: object
      additionalProperties: false
      required: [messageId, status, clientRequestId]
      properties:
        messageId:
          type: string
          format: uuid
        status:
          type: string
          const: pending
        clientRequestId:
          type: string
          minLength: 1
          maxLength: 255
...
    Attachment:
      type: object
      additionalProperties: false
      required: [id, fileName, contentType, sizeBytes, status]
      properties:
        id:
          type: string
          format: uuid
        fileName:
          type: string
          minLength: 1
          maxLength: 500
        contentType:
          type: string
          enum: [image/jpeg, image/png, video/mp4, application/pdf, application/octet-stream]
        sizeBytes:
          type: integer
          format: int64
          minimum: 0
          maximum: 20971520
        status:
          type: string
          enum: [pending, ready, failed, deleted]
    EmailSendMultipartRequest:
      type: object
      additionalProperties: false
      description: Multipart fields are ordered as to, subject, body, then zero or more repeated file parts. Attachment bytes are streamed and the aggregate file content is limited to 20 MiB.
      required: [to, subject]
      properties:
        to:
          type: string
          minLength: 1
          maxLength: 2048
          description: One recipient address encoded as UTF-8 text.
        subject:
          type: string
          maxLength: 998
          description: UTF-8 plain-text subject.
        body:
          type: string
          maxLength: 1048576
          description: Optional UTF-8 plain-text body; HTML and CID inline content are not supported.
        file:
          type: array
          maxItems: 16
          x-max-total-bytes: 20971520
          description: Repeated ordinary attachment parts, in user-selected order. Total content is at most 20 MiB.
          items:
            type: string
            format: binary
            maxLength: 20971520
    EmailSendJsonRequest:
      type: object
      additionalProperties: false
      required: [to]
      properties:
        to:
          type: string
          minLength: 1
          maxLength: 2048
        subject:
          type: string
          maxLength: 998
        body:
          type: string
          maxLength: 1048576
    EmailSendResult:
      type: object
      additionalProperties: false
      required: [messageId, from, to, subject, status]
      properties:
        messageId:
          type: string
          minLength: 1
          maxLength: 512
        from:
          type: string
          minLength: 1
          maxLength: 512
        to:
          type: string
          minLength: 1
          maxLength: 2048
        subject:
          type: string
          maxLength: 998
        status:
          type: string
          const: sent
...
    MessageKind:
      type: string
      enum: [text, template, image, video, document, email, system]
    MessageStatus:
      type: string
      enum: [pending, processing, submission_unknown, submitted, sent, delivered,
        read, failed, cancelled]